Add files via upload
A Program through which we can check about the drowsiness of the driver and raise an alarm accordingly. If the Euclidean distance goes below the defined limit, then the alarm is raised and the driver can get alert. 
This kind of program can be integrated into a Raspberry Pi device and can be implemented into the trucks/cars and can save numerous lives and avoid accidents.
Please check and contribute. All suggestions are welcome.